Bump version to v0.3.0-rc.1
Should contain the last major set of changes to the API design.
The change in the maning of DirectIdSet::remove is very severe breakage.
I will yank all the old pre-releases of idmap to ensure no one is relying on it.

Fix invalid cargo keywords.

It would be nice if we didn't have to comment out the dev-dependencies of idmap-derive
to publish things....
It would also be nice if `cargo package` had a warning for invalid keywords/categories.