derive: Fix expander conflicts between unit & integration tests
Need to add a random number to disambiguate between them,
as just using the builtin hash causes a conflict.
TODO: I should really open an issue for this on the `expander` repo.

Make expander conditional on a cfg(...) flag as well as a feature.

Adjust the Justfile to use --all-features on nightly.